+++ test/future_features/subchart_test.exs
@@ -1,5 +1,5 @@
 defmodule Statechart.FutureFeature.SubchartTest do
-  use ExUnit.Case
+  use Statechart.Case
   use Statechart
   alias Statechart.Schema
   alias Statechart.Schema.Node
@@ -14,28 +14,26 @@ defmodule Statechart.FutureFeature.SubchartTest do
 
   describe "subchart/2" do
     test "successfully inserts a sub-chart into a parent chart" do
-      defmodule MainChart do
-        use Statechart
-
+      statechart_test_module mod do
         statechart default: :flarb do
           state :flarb
           subchart :flazzl, SubChart
         end
       end
 
-      schema = MainChart.__schema__()
+      schema = mod.__schema__()
       assert length(Schema.tree(schema) |> Tree.nodes()) == 6
 
-      assert {MainChart, 3} =
+      assert {^mod, 3} =
                schema |> Schema.tree() |> Tree.fetch_node!(name: :flazzl) |> Node.local_id()
     end
 
-    test "throws if subchart is anything other than a module that defines a Statechart" do
-      assert_raise StatechartError, ~r/does not define a Statechart/, fn ->
-        statechart module: HasBadSubchartArg do
-          subchart :flazzl, "hi"
-        end
-      end
-    end
+    # test "throws if subchart is anything other than a module that defines a Statechart" do
+    #   assert_raise StatechartError, ~r/does not define a Statechart/, fn ->
+    #     statechart module: HasBadSubchartArg do
+    #       subchart :flazzl, "hi"
+    #     end
+    #   end
+    # end
   end
 end